+++ examples/performance-test/pom.xml
@@ -4,19 +4,18 @@
   <modelVersion>4.0.0</modelVersion>
 
   <parent>
-    <groupId>io.holixon.selectivereplay</groupId>
-    <artifactId>axon-selective-replay-examples-root</artifactId>
+    <groupId>io.holixon.axon</groupId>
+    <artifactId>axon-adhoc-projection-examples-root</artifactId>
     <version>0.0.1-SNAPSHOT</version>
   </parent>
 
-  <artifactId>axon-selective-replay-examples-performance-test</artifactId>
-  <name>${project.artifactId}</name>
-  <description>The core extension module</description>
+  <artifactId>axon-adhoc-projection-examples-performance-test</artifactId>
+  <description>Performance test</description>
 
   <dependencies>
     <dependency>
-      <groupId>io.holixon.selectivereplay</groupId>
-      <artifactId>axon-selective-replay-core</artifactId>
+      <groupId>io.holixon.axon</groupId>
+      <artifactId>axon-adhoc-projection-core</artifactId>
       <version>${project.version}</version>
     </dependency>
