+++ examples/pom.xml
@@ -4,13 +4,12 @@
   <modelVersion>4.0.0</modelVersion>
 
   <parent>
-    <groupId>io.holixon.selectivereplay</groupId>
-    <artifactId>axon-selective-replay-root</artifactId>
+    <groupId>io.holixon.axon</groupId>
+    <artifactId>axon-adhoc-projection-root</artifactId>
     <version>0.0.1-SNAPSHOT</version>
   </parent>
 
-  <artifactId>axon-selective-replay-examples-root</artifactId>
-  <name>${project.artifactId}</name>
+  <artifactId>axon-adhoc-projection-examples-root</artifactId>
   <description>Examples for using the library</description>
   <packaging>pom</packaging>
 
@@ -31,21 +30,6 @@
 
   <build>
     <plugins>
-      <plugin>
-        <groupId>com.tngtech.jgiven</groupId>
-        <artifactId>jgiven-maven-plugin</artifactId>
-        <version>${jgiven.version}</version>
-        <executions>
-          <execution>
-            <goals>
-              <goal>report</goal>
-            </goals>
-          </execution>
-        </executions>
-        <configuration>
-          <format>html</format>
-        </configuration>
-      </plugin>
       <plugin>
         <groupId>org.jacoco</groupId>
         <artifactId>jacoco-maven-plugin</artifactId>
