+++ extension/core/pom.xml
@@ -4,23 +4,24 @@
   <modelVersion>4.0.0</modelVersion>
 
   <parent>
-    <groupId>io.holixon.selectivereplay</groupId>
-    <artifactId>axon-selective-replay-extension</artifactId>
+    <groupId>io.holixon.axon</groupId>
+    <artifactId>axon-adhoc-projection-extension</artifactId>
     <version>0.0.1-SNAPSHOT</version>
   </parent>
 
-  <artifactId>axon-selective-replay-core</artifactId>
-  <name>${project.artifactId}</name>
+  <artifactId>axon-adhoc-projection-core</artifactId>
   <description>The core extension module</description>
 
   <dependencies>
     <dependency>
       <groupId>org.axonframework</groupId>
       <artifactId>axon-modelling</artifactId>
+      <scope>provided</scope>
     </dependency>
     <dependency>
       <groupId>org.axonframework</groupId>
       <artifactId>axon-messaging</artifactId>
+      <scope>provided</scope>
     </dependency>
     <dependency>
       <groupId>org.axonframework</groupId>
@@ -60,10 +61,10 @@
       <version>${mockk.version}</version>
       <scope>test</scope>
     </dependency>
+
     <dependency>
       <groupId>org.assertj</groupId>
       <artifactId>assertj-core</artifactId>
-      <version>${assertj.version}</version>
       <scope>test</scope>
     </dependency>
     <dependency>
