+++ extension/core/src/main/kotlin/io/holixon/axon/projection/adhoc/ModelFactory.kt
@@ -1,4 +1,4 @@
-package io.holixon.selectivereplay
+package io.holixon.axon.projection.adhoc
 
 import org.axonframework.eventhandling.DomainEventMessage
 import org.axonframework.eventsourcing.eventstore.DomainEventStream
