+++ extension/core/src/main/kotlin/io/holixon/axon/projection/adhoc/ModelInspector.kt
@@ -1,4 +1,4 @@
-package io.holixon.selectivereplay
+package io.holixon.axon.projection.adhoc
 
 import org.axonframework.messaging.annotation.*
 import java.lang.reflect.Constructor
