+++ extension/core/src/main/kotlin/io/holixon/axon/projection/adhoc/ModelRepository.kt
@@ -1,4 +1,4 @@
-package io.holixon.selectivereplay
+package io.holixon.axon.projection.adhoc
 
 import mu.KLogging
 import org.axonframework.common.caching.Cache
