+++ extension/core/src/test/kotlin/io/holixon/axon/projection/adhoc/dummy/events.kt
@@ -1,4 +1,4 @@
-package io.holixon.selectivereplay.dummy
+package io.holixon.axon.projection.adhoc.dummy
 
 import java.util.*
