+++ extension/core/src/test/resources/banner.txt
@@ -1,0 +1,3 @@
+-------------------------------------------------------------------------------
+I-Test, running Spring Boot ${spring-boot.version}
+-------------------------------------------------------------------------------
