+++ pom.xml
@@ -3,9 +3,10 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>io.holixon.selectivereplay</groupId>
-  <artifactId>axon-selective-replay-root</artifactId>
-  <name>axon-selective-replay-root</name>
+  <groupId>io.holixon.axon</groupId>
+  <artifactId>axon-adhoc-projection-root</artifactId>
+  <name>${project.artifactId}</name>
+  <description>This library provides a stateless model repository for Axon Framework using selective event loading executed during query.</description>
   <version>0.0.1-SNAPSHOT</version>
   <packaging>pom</packaging>
 
@@ -16,30 +17,19 @@
 
   <properties>
     <!-- JVM -->
-    <kotlin.compiler.jvmTarget>17</kotlin.compiler.jvmTarget>
     <java.version>17</java.version>
     <kotlin.version>1.9.10</kotlin.version>
 
     <!-- LIBS -->
-    <axon-bom.version>4.7.4</axon-bom.version>
-    <jgiven.version>1.3.0</jgiven.version>
-    <jgiven-kotlin.version>1.3.0.0</jgiven-kotlin.version>
-    <spring-boot.version>3.1.4</spring-boot.version>
     <kotlin-logging.version>3.0.5</kotlin-logging.version>
 
     <!-- TEST -->
-    <jackson.version>2.15.2</jackson.version>
-    <junit.version>4.13.2</junit.version>
-    <junit5.version>5.10.0</junit5.version>
     <dom4j.version>2.1.4</dom4j.version>
     <jaxen.version>2.0.0</jaxen.version>
-    <assertj.version>3.24.2</assertj.version>
     <mockk.version>1.13.8</mockk.version>
 
     <!-- MAVEN -->
-    <project.encoding>UTF-8</project.encoding>
-    <project.build.sourceEncoding>${project.encoding}</project.build.sourceEncoding>
-    <project.reporting.outputEncoding>${project.encoding}</project.reporting.outputEncoding>
+    <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
 
     <!-- deactivate by default -->
     <jacoco.skip>true</jacoco.skip>
@@ -65,7 +55,7 @@
       <dependency>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-dependencies</artifactId>
-        <version>${spring-boot.version}</version>
+        <version>3.1.4</version>
         <type>pom</type>
         <scope>import</scope>
       </dependency>
@@ -74,7 +64,7 @@
       <dependency>
         <groupId>org.axonframework</groupId>
         <artifactId>axon-bom</artifactId>
-        <version>${axon-bom.version}</version>
+        <version>4.8.2</version>
         <type>pom</type>
         <scope>import</scope>
       </dependency>
@@ -91,21 +81,6 @@
             <artifactId>kotlin-stdlib</artifactId>
           </exclusion>
         </exclusions>
-      </dependency>
-
-      <dependency>
-        <groupId>org.mockito.kotlin</groupId>
-        <artifactId>mockito-kotlin</artifactId>
-        <version>${mockito-kotlin.version}</version>
-        <scope>test</scope>
-      </dependency>
-
-      <dependency>
-        <groupId>com.fasterxml.jackson</groupId>
-        <artifactId>jackson-bom</artifactId>
-        <version>${jackson.version}</version>
-        <scope>import</scope>
-        <type>pom</type>
       </dependency>
 
     </dependencies>
@@ -158,9 +133,6 @@
               <plugin>spring</plugin>
               <plugin>all-open</plugin>
             </compilerPlugins>
-            <pluginOptions>
-              <option>all-open:annotation=io.holixon.testing.jgiven.AxonJGivenStage</option>
-            </pluginOptions>
           </configuration>
           <executions>
             <execution>
@@ -172,7 +144,6 @@
               <configuration>
                 <sourceDirs>
                   <sourceDir>${project.basedir}/src/main/kotlin</sourceDir>
-                  <sourceDir>${project.basedir}/src/main/java</sourceDir>
                 </sourceDirs>
               </configuration>
             </execution>
@@ -185,7 +156,6 @@
               <configuration>
                 <sourceDirs>
                   <sourceDir>${project.basedir}/src/test/kotlin</sourceDir>
-                  <sourceDir>${project.basedir}/src/test/java</sourceDir>
                 </sourceDirs>
               </configuration>
             </execution>
@@ -204,7 +174,7 @@
           <artifactId>maven-compiler-plugin</artifactId>
           <version>3.11.0</version>
           <configuration>
-            <encoding>${project.build.sourceEncoding}</encoding>
+            <encoding>UTF-8</encoding>
             <source>${java.version}</source>
             <target>${java.version}</target>
             <release>${java.version}</release>
@@ -329,22 +299,6 @@
           </configuration>
         </plugin>
 
-        <plugin>
-          <groupId>com.tngtech.jgiven</groupId>
-          <artifactId>jgiven-maven-plugin</artifactId>
-          <version>${jgiven.version}</version>
-          <executions>
-            <execution>
-              <goals>
-                <goal>report</goal>
-              </goals>
-            </execution>
-          </executions>
-          <configuration>
-            <format>html</format>
-          </configuration>
-        </plugin>
-
         <!-- Failsafe -->
         <plugin>
           <groupId>org.apache.maven.plugins</groupId>
@@ -364,7 +318,7 @@
             <sourceLinks>
               <link>
                 <path>${project.basedir}</path>
-                <url>https://github.com/holixon/axon-selective-replay/blob/master</url>
+                <url>https://github.com/holixon/axon-adhoc-projection/blob/master</url>
                 <lineSuffix>#L</lineSuffix>
               </link>
             </sourceLinks>
@@ -606,7 +560,7 @@
   <distributionManagement>
     <site>
       <id>README</id>
-      <url>https://github.com/holixon/axon-selective-replay</url>
+      <url>https://github.com/holixon/axon-adhoc-projection</url>
     </site>
     <snapshotRepository>
       <id>ossrh</id>
@@ -621,14 +575,14 @@
   </distributionManagement>
 
   <scm>
-    <connection>scm:git:git://github.com/holixon/axon-selective-replay.git</connection>
-    <developerConnection>scm:git:git@github.com:holixon/axon-selective-replay.git</developerConnection>
-    <url>https://github.com/holixon/axon-selective-replay</url>
+    <connection>scm:git:git://github.com/holixon/axon-adhoc-projection.git</connection>
+    <developerConnection>scm:git:git@github.com:holixon/axon-adhoc-projection.git</developerConnection>
+    <url>https://github.com/holixon/axon-adhoc-projection</url>
     <tag>HEAD</tag>
   </scm>
   <issueManagement>
     <system>GitHub</system>
-    <url>https://github.com/holixon/axon-selective-replay/issues</url>
+    <url>https://github.com/holixon/axon-adhoc-projection/issues</url>
   </issueManagement>
 
   <developers>
